+++ packages/telescope-daily/lib/daily.js
@@ -6,4 +6,4 @@ Telescope.menuItems.add("viewsMenu", {
   description: 'day_by_day_view'
 });
 
-_.findWhere(Telescope.modules.top,{template: "posts_views_nav"}).only.push("postsDaily");
+Telescope.modules.addRoute("top", "views_menu", "postsDaily");