+++ packages/telescope-singleday/lib/singleday.js
@@ -18,4 +18,4 @@ Posts.views.add("singleday", function (terms) {
   };
 });
 
-_.findWhere(Telescope.modules.top,{template: "posts_views_nav"}).only.push("postsSingleDay");
+Telescope.modules.addRoute("top", "views_menu", "postsSingleDay");